+++ backend/app/controllers/spree/admin/base_controller.rb
@@ -3,6 +3,8 @@
 module Spree
   module Admin
     class BaseController < Spree::BaseController
+      include Spree::Admin::Breadcrumbs
+
       helper 'spree/admin/navigation'
       layout '/spree/layouts/admin'
 
@@ -61,28 +63,6 @@ module Spree
 
       def order_mutex_redirect_path
         edit_admin_order_path(@order)
-      end
-
-      def add_breadcrumb(name, path = nil)
-        @admin_breadcrumbs ||= []
-        @admin_breadcrumbs << [name, path]
-      end
-
-      # Shared breadcrumbs
-
-      def set_user_breadcrumbs
-        add_breadcrumb plural_resource_name(Spree::LegacyUser), spree.admin_users_path
-        add_breadcrumb @user.email, edit_admin_user_url(@user) if @user && !@user.new_record?
-      end
-
-      def set_order_breadcrumbs
-        add_breadcrumb plural_resource_name(Spree::Order), spree.admin_orders_path
-        add_breadcrumb "##{@order.number}", spree.edit_admin_order_path(@order) if @order && !@order.new_record?
-      end
-
-      def set_product_breadcrumbs
-        add_breadcrumb plural_resource_name(Spree::Product), spree.admin_products_path
-        add_breadcrumb @product.name, spree.admin_product_path(@product) if @product && !@product.new_record?
       end
     end
   end
